Refactor react blocks and table messages
