+++ client/src/redux/actions/game/gameAction.js
@@ -1,10 +1,10 @@
 import * as types from "../../types/game/gameType";
-import * as typesGame from "../../types/game/gameType";
+import * as typesMessage from "../../types/game/messageType";
 
 export const setPlayer = (currentPlayer) =>async dispatch=>{
     localStorage.setItem('currentPlayer', JSON.stringify(currentPlayer));
     dispatch({
-        type: typesGame.GAME_SET_PLAYER,
+        type: types.GAME_SET_PLAYER,
         player: currentPlayer,
     });
 }
@@ -18,27 +18,27 @@ export const setCurrentRound = (currentRound) =>async dispatch=>{
 
 export const setChat = (round) =>async dispatch=>{
     dispatch({
-        type: typesGame.GAME_SET_CHAT_ALL,
+        type: types.GAME_SET_CHAT_ALL,
         chat: round.messages,
     });
 }
 export const clearChat = () =>async dispatch=>{
     dispatch({
-        type: typesGame.GAME_SET_CHAT_ALL,
+        type: types.GAME_SET_CHAT_ALL,
         chat: [],
     });
 }
 
 export const setGame = (game) =>async dispatch=>{
     dispatch({
-        type: typesGame.GAME_SET_GAME,
+        type: types.GAME_SET_GAME,
         game: game,
     });
 }
 
-export const setMessage = (key, message) =>async dispatch=> {
+export const setTableMessage = (message) =>async dispatch=> {
     dispatch({
-        type: typesGame.GAME_SET_CHAT_MESSAGE,
-        message: {_id: key, text: message, type: 'system'},
+        type: typesMessage.MESSAGE_SET_TABLE_MESSAGE,
+        text: message,
     });
 }
