+++ client/src/redux/types/game/messageType.js
@@ -1,1 +1,2 @@
-export const MESSAGE_CHANGE_TEXT = 'MESSAGE_CHANGE_TEXT';
+export const MESSAGE_CHANGE_TEXT_MESSAGE = 'MESSAGE_CHANGE_TEXT_MESSAGE';
+export const MESSAGE_SET_TABLE_MESSAGE = 'MESSAGE_SET_TABLE_MESSAGE';
